updates read me and refactors code
