+++ README.md
@@ -2,21 +2,147 @@ To easily add ads into your app.
 
 ## Features
 
-TODO: List what your package can do. Maybe include images, gifs, or videos.
+- Google Mobile Ads (banner, interstitial, rewarded video ad)
 
-## Getting started
+## Prerequisites
 
-TODO: List prerequisites and provide or point to information on how to
-start using the package.
+*   Flutter 1.22.0 or higher
+*   Android
+    *   Android Studio 3.2 or higher
+    *   Target Android API level 19 or higher
+    *   Set `compileSdkVersion` to 28 or higher
+    *   Android Gradle Plugin 4.1 or higher (this is the version supported by Flutter out of the box)
+*   Ios
+    *   Latest version of Xcode with [enabled command-line tools](https://flutter.dev/docs/get-started/install/macos#install-xcode).
+*   Recommended: [Create an AdMob account](https://support.google.com/admob/answer/2784575) and [register an Android and/or iOS app](https://support.google.com/admob/answer/2773509) (To show live ads on a published app, it is required to register that app).
 
-## Usage
+## Platform Specific Setup
 
-1- Call this method at start
-```dart
-await EasyAds.instance.initialize();
+
+### iOS
+
+
+#### Update your Info.plist
+
+Update your app's `ios/Runner/Info.plist` file to add two keys:
+
+* A `GADApplicationIdentifier` key with a string value of your AdMob app ID ([identified in the AdMob UI](https://support.google.com/admob/answer/7356431)).
+* A `SKAdNetworkItems` key with Google's `SKAdNetworkIdentifier` value of `cstr6suwn9.skadnetwork`.
+
+```xml
+<key>GADApplicationIdentifier</key>
+<string>ca-app-pub-3940256099942544~1458002511</string>
+<key>SKAdNetworkItems</key>
+  <array>
+    <dict>
+      <key>SKAdNetworkIdentifier</key>
+      <string>cstr6suwn9.skadnetwork</string>
+    </dict>
+  </array>
 ```
 
-## Additional information
+See https://developers.google.com/admob/ios/quick-start#update\_your\_infoplist for more information about configuring `Info.plist` and setting up your App ID.
 
-For more info:
 
+### Android
+
+
+#### Update AndroidManifest.xml
+
+The AdMob App ID must be included in the `AndroidManifest.xml`. Failure to do so will result in a crash on launch of an app.
+
+Add the AdMob App ID ([identified in the AdMob UI](https://support.google.com/admob/answer/7356431)) to the app's `android/app/src/main/AndroidManifest.xml` file by adding a `<meta-data>` tag with name `com.google.android.gms.ads.APPLICATION_ID`, as shown below. You can find your App ID in the AdMob UI. For `android:value` insert your own AdMob App ID in quotes, as shown below.
+
+
+```xml
+<manifest>
+    <application>
+        <!-- Sample AdMob App ID: ca-app-pub-3940256099942544~3347511713 -->
+        <meta-data
+            android:name="com.google.android.gms.ads.APPLICATION_ID"
+            android:value="ca-app-pub-xxxxxxxxxxxxxxxx~yyyyyyyyyy"/>
+    </application>
+</manifest>
+```
+
+
+The same value when you initialize the plugin in your Dart code.
+
+See https://goo.gl/fQ2neu for more information about configuring `AndroidManifest.xml` and setting up the App ID.
+
+
+## Initialize the SDK
+
+Before loading ads, have your app initialize the Mobile Ads SDK by calling `EasyAds.instance.initialize()` which initializes the SDK and returns a `Future` that finishes once initialization is complete (or after a 30-second timeout). This needs to be done only once, ideally right before running the app.
+
+
+```dart
+import 'package:easy_ads_flutter/easy_ads_flutter.dart';
+import 'package:flutter/material.dart';
+
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await EasyAds.instance.initialize();
+
+  runApp(MyApp());
+}
+
+class MyApp extends StatefulWidget {
+  @override
+  MyAppState createState() => MyAppState();
+}
+
+class MyAppState extends State<MyApp> {
+  @override
+  void initState() {
+    super.initState();
+    // Load ads.
+  }
+}
+```
+
+## Initialize the admob
+
+```dart
+import 'package:easy_ads_flutter/easy_ads_flutter.dart';
+
+EasyAds.instance.initAdmob(
+        interstitialAdUnitId: InterstitialAd.testAdUnitId,
+        rewardedAdUnitId: RewardedAd.testAdUnitId);
+```
+
+## load ad
+
+```dart
+import 'package:easy_ads_flutter/easy_ads_flutter.dart';
+
+EasyAds.instance.loadRewardedAd();
+
+OR
+
+EasyAds.instance.loadInterstitialAd();
+```
+
+## showing ad
+
+```dart
+import 'package:easy_ads_flutter/easy_ads_flutter.dart';
+
+EasyAds.instance.showInterstitialAd();
+
+OR
+
+EasyAds.instance.showRewardedAd();
+```
+
+## dispose ad
+
+```dart
+import 'package:easy_ads_flutter/easy_ads_flutter.dart';
+
+EasyAds.instance.disposeInterstitialAd();
+
+OR
+
+EasyAds.instance.disposeRewardedAd();
+```
