+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: easy_ads_flutter
 description: Easy Ads is a wrapper around famous ad packages which let you integrate ads easily
-version: 0.0.1
+version: 0.0.2
 homepage: https://nooralibutt.com/
 repository: https://github.com/nooralibutt/easy-ads
